minor change: don't update an empty list of items
git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@14080 1dfba125-df10-0410-8a77-af1a8d6d074f

+++ bin/upgrading/texttool.pl
@@ -673,7 +673,7 @@ sub remove {
         $txtids .= $txtid;
     }
     $dbh->do("DELETE FROM ml_latest WHERE dmid=$dmid AND itid=$itid");
-    $dbh->do("DELETE FROM ml_text WHERE dmid=$dmid AND txtid IN ($txtids)");
+    $dbh->do("DELETE FROM ml_text WHERE dmid=$dmid AND txtid IN ($txtids)") if $txtids;
 
     $out->("-","done.");
 }
